* Switched the SDK from checked-in projects to VPC, the Valve Project Creator. See the Getting Started document on the wiki for details. * Pulled in bug fixes from HL2 and HL2MP.

+++ mp/src/public/dt_common.h
@@ -190,7 +190,7 @@ public:
 	union
 	{
 		float	m_Float;
-		long	m_Int;
+		int		m_Int;
 		const char	*m_pString;
 		void	*m_pData;	// For DataTables.
 #if 0 // We can't ship this since it changes the size of DTVariant to be 20 bytes instead of 16 and that breaks MODs!!!
